Graceful fallback for nvim compiled without luajit
Lua proper does not provide the jit helper variable.

Nvim can be compiled with liblua5.1 in place of libluaji5.1. Luajit is
not available on all architectures.

This will make the Unix features unavailable even if the target platform
is Linux/MacOS when running under nvim compiled against liblua5.1, but
the plugin will not error out.