fix: Ignore section headings in md codeblocks
This fix is related to issue #145. A new utility function is added which
checks whether the cursor is currently inside of a code block by
counting the instances of backticks before or after the cursor in the
buffer. If the cursor is in a code block, any headings (or things that
resemble headings, such as comments introduced by #) are ignored.