fix: Avoid re-folding; use existing folds
This commit changes the way that the folding function works. Instead of
always creating a new fold (which is what the function previously did),
it now looks to see if the cursor is already in an open fold, and if so,
it will simply close the fold. If there is no fold, one will be created
following the usual logic.