refactor: Split to-do functionality into to_do.lua
This commit does a major refactor of the to-do list management
functionality that was previously in the `lists` module. All
to-do-list-specific functionality has now been split off into a separate
module, `to_do.lua` The new module is a significant improvement in the
implementation of the functionality and also uses LuaCATS annotations
for improved documentation.